[MM-32476] Apps in the Marketplace (#1352)
